+++ hutool-http/src/main/java/cn/hutool/http/useragent/UserAgentUtil.java
@@ -18,5 +18,4 @@ public class UserAgentUtil {
 		return UserAgentParser.parse(userAgentString);
 	}
 
-
 }